+++ src/lib.rs
@@ -20,8 +20,6 @@ pub mod config_tester;
 mod config_tester_profiles;
 mod config_tester_runtime;
 mod config_tester_upstreams;
-#[cfg(feature = "ingress")]
-mod http_types;
 pub mod internal_crypto;
 #[cfg(feature = "metrics")]
 pub mod metrics;